Moved gui files to lower level directory
Moved gui files to lower level directory. Fixed import to reflect move.
